Update team_join and seed new users with on boarding flow
