+++ src/bot/lib/constants.js
@@ -617,6 +617,14 @@ export const buttonValues ={
 	letsDoIt: {
 		name: "LETS_DO_IT",
 		value: "lets do it!"
+	},
+	changeTime: {
+		name: "CHANGE_TIME",
+		value: "change time!"
+	},
+	keepTime: {
+		name: "KEEP_TIME",
+		value: "keep time!"
 	}
 }
